Convert write() "properties" argument to colon separated list.
Also added "properties_array" argument to allow the list of properties
to be specified as an array. An error occurs if both properties and
properies_array are present. Updated main programs and quippy so that
'make all' and 'python setup.py test' succeed, respectively.

At same time, error handling added to parse_string().

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@6908 9aeadc09-fa51-0410-b5d8-e5c65b1e449c