Improvements to write() when writing to an Extendable_str
Newlines are now added in Fortran not C, using add_lf_if_missing
argument to Extendable_str concat(). This means that choice of '^J'
vs. '^M' should be consistent with that used when reading from an
Extendable_str.

quippy gains a more Pythonic interface to the Extendable_str class,
with str(), len() and indexing.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7038 9aeadc09-fa51-0410-b5d8-e5c65b1e449c