Changes to top-level potential_calc() interface.
Long-discussed extension of the top level potential_calc() interface.
There are now 4 optional arguments: energy, force, virial, and
local_energy.  These behave the same as before.  However, now you can
also specify quantities to be computed in the calc args_str.  The
keywords are the same as the optional argument names.  If you just
specify "energy", the calculated energy will be put in a parameter
called energy (same for virial, and force/local_energy will be in
same-named per-atom properties).  You can also do
"energy=my_special_energy", and then it'll end up in a parameter
called my_special_energy.  Note that this applies to parsing of
strings in general.  A string argument specified with no "=VALUE" will
end up with a value of its own name (e.g. "energy=energy").

If you pass an optional argument and an args_str, the value will
be returned in both places (the argument and the parameter/property).

Changes to Dictionary.f95:

   Add set_value_none(this, key) to set_value() interface.

   Required so that read_string() and write_string() can be inverses
   of one another when reading a string containing only a key with no
   associated value, e.g. an args_str of the kind used by new
   Potential_calc() interface "energy force virial". The type is set
   to T_NONE.

   In future we may want to modify Dictionary read_string() so that
   it defers parsing of the values until get_value() is called, since
   then we have more information about the expected type.

quippy changes to args_str handling:

   If we're expecting an args_str argument, then any kwarg is
   permitted. Otherwise kwarg name must match one of the expected
   names. If kwarg name matches one of expected values, but type of
   argument does not, then we assume it's an arg_str argument.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7100 9aeadc09-fa51-0410-b5d8-e5c65b1e449c