Small bug fix to IPModel_GAP calc() for case of do_lammps = .false.
ForceMixing Potential now with atom_mask as well as with cluster
carving. Not very efficient when r_scale != 1, since we rescale the
whole system, and we have to call calc_dists() to update neighbour
distances. Only the active atoms are included in atom_mask, not buffer
are marked: current GAP atom_mask implementation takes care of
buffering automatically.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7622 9aeadc09-fa51-0410-b5d8-e5c65b1e449c