Remove counter-intuitive transposition of FortranArray on initialisation
Previously, farray() and FortranArray() implicitly transposed input if
it was a nested list or tuple, but not if it was an ndarray or
FortranArray index. Now we no longer do this, since it's confusing and
unintuitive: instead we now do an explicit tranposition if it's
necessary.

The FortranArray.transpose_on_print attribute has also been
removed. Use, e.g. 'print at.pos.T' for a nicely formatted printout of
the atomic positions.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7657 9aeadc09-fa51-0410-b5d8-e5c65b1e449c