When writing NetCDF files, rotate atomic positions to match cell orientation
In a NetCDF file following th AMBER conventions, the 'a' cell vector
must be oriented along the x axis, and the 'b' cell vector must lie in
the x-y plane. We now follow these conventions correctly, if necessary
transforming the atomic positions to match this cell orientation. This
is done by transforming at%pos to match the rotated lattice in such a
way that the fractional positions (at%g .mult. at%pos) are preserved.

The CInOutput write() interface applies the transformation to at%pos
before calling the lowlevel C routine write_netcdf(). The original
lattice is stored in the NetCDF file as a 3x3 variable named
cell_lattice. If a rotation was necessary, the NetCDF variable
cell_rotation is set to 1. After writing the file, the lattice
and positions are restored to their original values before returning
from write().

When reading from a NetCDF file with the read() interface, we check if
the cell_rotation variable is defined. If not, we assume the cell
vectors are oriented according to the AMBER conventions, for backwards
compatibility with old files. If cell_rotation == 1, we transform the
atomic positions to match the original stored lattice in the
cell_lattice variable.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7658 9aeadc09-fa51-0410-b5d8-e5c65b1e449c