Added atoms_transform_basis() and atoms_rotate() routines
These routines transform rank 1 and rank 2 tensors in this%params and
this%properties. Tensors are identified by having the correct type
(real arrays) and shape (i.e. 3, (3, 3), (3, this%N) (9, this%N) for
vector paramters, tensor parameters, vector properties and tensor
properties respectively), and by having a name which is included in
the relevant list. Extra names can be added to the lists with the
rank1 and rank2 arguments.

The default rank1 and rank2 names are as follows:

  default_rank1 = "pos:velo:acc:avgpos:oldpos:force:efield:dipoles"
  default_rank2 = "virial:local_virial"

Simple tests are in test_atoms.TestAtoms_Rotate.

Also added wrapping of Quaternion type to quippy (useful for
generating rotation matrices), and is_orthonormal() test to
linearalgebra module.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@7678 9aeadc09-fa51-0410-b5d8-e5c65b1e449c