Allow multiple <md> stanzas in crack XML input file
Stanzas are exectuted in sequence, in the order given in input
file. Upto MAX_MD_STANZA (current value is 5) are allowed.

The duration of each stanza is set by a new parameter, stanza_time,
which should be a value in femtoseconds. Any MD parameters can be
given new values. When moving to a new stanza, all thermostats are
removed and recreated.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@8016 9aeadc09-fa51-0410-b5d8-e5c65b1e449c