Support for wrapping interfaces outside of classes in quippy
Top-level interfaces are now wrapped in the same way as those inside
classes. This means overloaded routines are available in quippy with
the Fortran generic interface name, and the correct specific routine
will be selected at runtime based on types and ranks of arguments.

Example:

>>> a = fzeros(10)
>>> randomise(a, 0.5) # calls array_randomise(a, 0.5)
>>> b = farray(10,10)
>>> randomise(b, 0.5) # calls matrix_randomise(b, 0.5)

The unit tests have been updated but there will be some scripts which
need updating, since the specific names are no longer available from
quippy.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@8512 9aeadc09-fa51-0410-b5d8-e5c65b1e449c