MPI bug fix to dictionary_bcast()
When initialising array entries, both entry%len (1D length) and
entry%len2 (2D length) must be set. Otherwise, add_entry() may
incorrectly overwrite an existing array with a new one if there is
a mismatch in either len or len2. A better fix would be not to store
len and len2 and explicitly test size of stored array.

git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@8519 9aeadc09-fa51-0410-b5d8-e5c65b1e449c