Added in_out_in_mode args_str option to create_cluster_info_from_hybrid_mark()
This option affects how IN-OUT-IN heuristic is applied. Default is
"all" which is same as old behaviour: ALL nearest neighbours of OUT
atom must be IN for it to be added. Other option is "any" which will
add atom if ANY of it's nearest neighbours (apart from the original IN
atom) are IN. Stronger requirement is needed for CP2K compatibility.



git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@9080 9aeadc09-fa51-0410-b5d8-e5c65b1e449c