calc_connect() keeps track of distance moved since last connect
Changes to calc_connect() which should speed up connectivity
calculations during MD and minimisations. The idea is that
calc_connect() now keeps track of how far atoms have moved since the
last full calculation of the connectivity, and in between cit simply
calls calc_dists() to update the stored distance tables.

Note that if the lattice changes then a full calc_connect() is still
performed, as I didn't figure out how to calculate minimum images
distances between a pair of configurations with difference lattices.

To enable this feature you need to set the Atoms cutoff_skin attribute
to a non-zero value. This is the amount by which the Atoms cutoff
should be increased to allow a buffer, and can be set with the
set_cutoff() routine, which has a new cutoff_skin optional
argument. The eval main program also has a cutoff_skin command line
argument, with default value 0.5 A.

These changes necessitated moving diff_min_image() and
distance_min_image() from the Atoms module to Atoms_types, and
calc_dists() from Atoms to Connection.

In tests on a 512,000 atom system with a 5.0 A cutoff, I get about a
factor of three speedup in the average time to update connectivity
during dynamics by using cutoff_skin=1.0 A (from 6.5s to 1.2s).