Additional status outputs from calc_connect()
Added two optional, intent(out) arguments to connection_calc_connect()
and atoms_calc_connect(): max_pos_change, which is a real number
giving the largest movement of any atom since the last connectivity
rebuild, and did_rebuild which is a logical that is set to .true. when
a rebuild is done.