Fixed memory leak when removing Dictionary entries
Albert spotted a memory leak that occurs when repeatedly
adding/removing properties to an Atoms object. This was caused by not
freeing the memory associated with the removed array, and has now been
fixed by adding 'call finalise(this%entry(this%n))' to
libAtoms/Dictionary.f95:2510.

Here's a program that shows the leak when run through valgrind:

[kermode@kasparov QUIP]$ cat QUIP_Programs/leak.f95
program leak

  use libatoms_module

  implicit none

  type(Atoms) :: at
  integer i

  call system_initialise
  trace_memory = .true.

  call diamond(at, 5.44_dp, (/ 14 /))

  do i=1,10000
     call add_property(at, 'tmp', 0)
     call remove_property(at, 'tmp')
  end do

  call system_finalise

end program leak