Bug fix to evaluation of Nye tensor.
Expression for \alpha_{kj} on line 127 of nye_tensor.f95 should be:

   \alpha_{jk} = -\epsilon{jim} \partial_i G_{mk}

instead of previous:

    \alpha_{jk} = -\epsilon{jim} \partial_k G_{im}

which follows Eq. 22 in the Hartley and Mishin (2005) Acta. Mat. paper, and
is presuambly a typo in original paper.

With the old code, I get nonzero alpha(1,1) and zero alpha(3,3)
entries for a screw dislocation, which seemed wrong. After applying
this patch I get much more sensible results.