Change default NetCDF output format from NetCDF4 to NetCDF3
Following suggestion from Lars:

"The problem is that HDF5 modifies the header of the database every
time it writes to the file. This means that data not yet written won't
take up space in the file, but it easily leads to data corruption.

If you use the native NetCDF database format, once the file is created
data corruption typically doesn't occur. That is because NetCDF creates
dummy data for entries that are not yet written and does not need to
reshape the arrangement of data in the file. My solution is to use
64-bit NetCDF files for the time being."