Option of legacy masses and units
Add two compile time flags (`DEFINES += ...` in `Makefile.inc`)

 -  `-DLEGACY_MASSES`
 -  `-DLEGACY_UNITS`

which revert the values to pre-2017 values, for code that depends on the
old values. Tests will fail if using these values, but keep them around
just in case. Can use as a framework for future updates of CODATA and
IUPAC.

In combination with the other commits, closes #54