Go back to manual API version for quip/lammps interface
This is so we can have the version number increment only in the case of
incompatible changes to the interface, which is what we really needed
(API versions need to exactly match for compatiblity; interface changes
are not likely to be forwards or backwards compatible).

This part-reverts commit 7d3de46dd0d881cc7084b68b2e6728f07714283c.