Improve NetCDF library config and linking
`make config` has been modified to ask for the compilation
and linking flags for NetCDF. It will default to the flags
returned from either `pkg-config` or `nc-config`, but allows
and flags to be entered freeform, so should work for any
installation (previously it assmued library names, tried to
read files that may not exist and and linked many unnecessary
libraries).

Also drops support for NetCDF3 (NetCDF4 can read/write version
3 files anyway).

Since flag is changed to `HAVE_NETCDF4`, `make config` can be
re-run without having to delete the previous `Makefile.inc`
(but will default to no NetCDF if `make config` is not re-run).