Remove adsorbate_info property in quippy
The adsorbate_info property in ase is long gone, but still exists
as a property to give a warning. Some parts of quippy, e.g. oo_fortran
call dir on Atoms to get all the attributes, which leads to errors
since it tries to directly access adsorbate_info, nobody actually uses
it. Getting rid of the property means it will just stay in the info
dict. This also fixes print for Atoms!