feat: decouple blocksizes by cheating pdtrtrs
While the ScaLAPACK routines pdgeqrf and pdormqr need large work arrays
for large nb, pdtrtrs requires mb == nb for unknown reasons. We only use
a single process column, so we can use an independent nb for the call of
pdtrtrs (3rd) without enforcing the restriction on the other procedures.
This saves both memory and computation time by not having to add zero
rows to have m be a multiple of nb.